[BE] EPMRDUAITM-136 Сreate room reading endpoint (#115)
* [BE] EPMRDUAITM-136 Added endpoint validation of room reading

* [BE] EPMRDUAITM-136 Added methods to RoomRepository to get room by codes

* [BE] EPMRDUAITM-136 Implement GetRoomHandler to get room from DB

* [BE] EPMRDUAITM-136 Implemented reading room endpoint

* [BE] EPMRDUAITM-136 Improved reading room endpoint documentation

* [BE] EPMRDUAITM-136 Added endpoint validation of room reading

* [BE] EPMRDUAITM-136 Added methods to RoomRepository to get room by codes

* [BE] EPMRDUAITM-136 Implement GetRoomHandler to get room from DB

* [BE] EPMRDUAITM-136 Implemented reading room endpoint

* [BE] EPMRDUAITM-136 Improved reading room endpoint documentation

* [BE] EPMRDUAITM-136 Small fixes

* [BE] EPMRDUAITM-136 Fix reading room request validation

* [BE] EPMRDUAITM-136 Simplified methods to get room by code

* [BE] EPMRDUAITM-136 Cleaned up repositories, removed unused methods and base repository

* [BE] EPMRDUAITM-136 Added IsFull property to the Room model

+++ backend/ApiService/Source/Domain/Builders/RoomBuilder.cs
@@ -10,9 +10,9 @@ namespace Epam.ItMarathon.ApiService.Domain.Builders
     {
         private DateTime? _closedOn;
         private string _invitationCode;
-        private ulong _minUsersLimit = 3;
-        private ulong _maxUsersLimit = 20;
-        private ulong _maxWishesLimit  = 5;
+        private uint _minUsersLimit = 3;
+        private uint _maxUsersLimit = 20;
+        private uint _maxWishesLimit  = 5;
         private string _name;
         private string _description;
         private string _invitationNote = "Hey!\nJoin our Secret Nick and make this holiday season magical! 🎄\nYou‘ll get to surprise someone with a gift — and receive one too. 🎅✨\nLet the holiday fun begin! 🌟 \n🎁 Join here:";
@@ -30,17 +30,17 @@ namespace Epam.ItMarathon.ApiService.Domain.Builders
             _invitationCode = invitationCode;
             return this;
         }
-        public RoomBuilder WithMinUsersLimit(ulong minUsersLimit)
+        public RoomBuilder WithMinUsersLimit(uint minUsersLimit)
         {
             _minUsersLimit = minUsersLimit;
             return this;
         }
-        public RoomBuilder WithMaxUsersLimit(ulong maxUsersLimit)
+        public RoomBuilder WithMaxUsersLimit(uint maxUsersLimit)
         {
             _maxUsersLimit = maxUsersLimit;
             return this;
         }
-        public RoomBuilder WithMaxWishesLimit(ulong maxWishesLimit)
+        public RoomBuilder WithMaxWishesLimit(uint maxWishesLimit)
         {
             _maxWishesLimit = maxWishesLimit;
             return this;
